Organized config code files
