+++ lua/lual/config.lua
@@ -1,29 +1,31 @@
---- Configuration API (Backwards Compatibility Wrapper)
--- This module provides backwards compatibility by delegating to the modular config system
+--- Configuration API (Main Entry Point)
+-- This module provides the main configuration API for the lual logging library
 
-local modular_config = require("lual.config.init")
+local api = require("lual.config.api")
 
 local M = {}
 
--- All the complex logic is now in the modular config system
--- This is just a delegation wrapper
-
---- Delegates to the modular config system
+--- Updates the _root logger configuration with the provided settings
 -- @param config_table table Configuration updates to apply
 -- @return table The updated _root logger configuration
 function M.config(config_table)
-    return modular_config.config(config_table)
+    return api.config(config_table)
 end
 
---- Delegates to the modular config system
+--- Gets the current _root logger configuration
 -- @return table A copy of the current _root logger configuration
 function M.get_config()
-    return modular_config.get_config()
+    return api.get_config()
 end
 
---- Delegates to the modular config system
+--- Resets the _root logger configuration to defaults
 function M.reset_config()
-    return modular_config.reset_config()
+    return api.reset_config()
+end
+
+--- Reset registry (for testing)
+function M.reset_registry()
+    return api.reset_registry()
 end
 
 return M
